patch

+++ IoT_Terminal/app/src/main/java/com/hit_src/iot_terminal/object/sensortype/op/Add.java
@@ -6,7 +6,7 @@ public class Add extends OP {
     }
 
     @Override
-    public int calculate(int reg, int[] source) {
+    public int calculate(int reg, byte[] source) {
         if(isMemory){
             return reg+source[value];
         } else{
